move bot logic from main yo new bot module
